app/requirements.txt
appsignal
opentelemetry-instrumentation-flask
Flask
